V1
Had a couple issues, mainly with the homepage and footer. Will definitely try and make them a bit prettier and complete. I could not get the fonts you wanted and not having the images or your logo made it a bit trickier and make the pages look less visually appealing. Excited to revise and enhance this